Add Gradio demo and fix GPU numerical accuracy on Ampere+ GPUs
The FRE constrained convolution produced divergent results on A100 GPUs
due to cuDNN's TF32 mode (10-bit mantissa), with per-pixel errors of ~0.1
amplified ~800x through the constrained least-squares solver, causing
z-score differences up to 387. Fixed by running the full FRE+FSD pipeline
in float64 and disabling TF32. CPU/GPU results now match to <5e-9.

Changes:
- fsd/fre.py: Constrain weights in the input's dtype (float64 when given
  float64 input) so GPU constraint math matches CPU exactly
- fsd/fsd_computation.py: Create image tensors as float64 instead of float32
- fsd/detector.py: Disable TF32 at model load time as safety net
- demo.py: Interactive Gradio web demo with dark theme, probability display,
  gauge visualization, and detailed result explanations
- pyproject.toml: Add gradio dependency
- README.md: Document the Gradio demo

+++ fsd/fsd_computation.py
@@ -94,9 +94,9 @@ def compute_fsd(
     if isinstance(image, Image.Image):
         if image.mode != "L":
             image = image.convert("L")
-        image_t = torch.from_numpy(__import__("numpy").array(image)).float().unsqueeze(0)  # (1, H, W)
+        image_t = torch.from_numpy(__import__("numpy").array(image)).double().unsqueeze(0)  # (1, H, W)
     elif isinstance(image, torch.Tensor):
-        image_t = image.float()
+        image_t = image.double()
         if image_t.ndim == 2:
             image_t = image_t.unsqueeze(0)
     else:
